fix(cli): set no_confirm=True when switching to non-interactive mode
When stdin is not a TTY and prompts are provided, we switch to
non-interactive mode but weren't setting no_confirm=True. This caused
the code to try to get user confirmation in non-interactive mode,
leading to error 19 'Operation not supported by device' when trying
to do terminal operations.

This bug was exposed when testing shell commands with pipes and stderr
redirects, as they would fail with error 19 unless -y flag was used.